+++ slacky/__main__.py
@@ -37,7 +37,8 @@ commands = {
         # Uncomment to load deepfry, requires OpenCV 4
         # 'deepfry': lambda  **payload: deepfry,
         # Uncomment to load shutterstock, requires API credentials,
-        'stockpic': lambda **payload: stockpic
+        'stockpic': lambda **payload: stockpic,
+        'coronastatus': lambda **payload: coronavirus
     }
 
 @slack.RTMClient.run_on(event='message')